requirements.txt
blinker==1.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
docutils==0.21.2
filetype==1.2.0
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Kivy==2.3.1
Kivy-Garden==0.1.5
kivy_deps.angle==0.4.0
kivy_deps.glew==0.3.1
kivy_deps.sdl2==0.8.0
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
Pygments==2.19.2
pypiwin32==223
python-dotenv==1.1.1
pywin32==310
requests==2.32.4
urllib3==2.5.0
Werkzeug==3.1.3
